requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
reportlab
urllib3
deepeval
gunicorn